DOM Caching Feature: Cache header to improve initial response
This feature caches the header portion of a webpage
when the user visits the webpage for the 1st time.
On subsequent visits to the page, we paint the header
instantly before the page is fetched over the network.
This improves user's initial response.

Change-Id: I66e89d1f222baf131e4d2b0b1fa2bf855ab62e7b

DOM Caching: Fix to check user agent before painting header

We now save the user agent associated with a cached header
and check the user agent against the current url being
loaded before painting the header.

This insures that we do not paint mobile header when the user has switched
the user agent to desktop and vise versa.

Change-Id: Ie35b633d54fafc366103ca53ea11ff0702698b54

Null check for elements while traversing DOM tree in Browser

Browser traverses the whole dom tree to find all link elements needed
for dns prefetch. No null check for received element pointer was
causing browser to crash.

Change-Id: I374de638a896c671dd8b1098f1afde60b0421b6d