Implement all WebKit logging functions on Android
Currently, all WebKit logging goes to stderr. This change hooks it up to
logcat. We use a single log level (LOG_DEBUG) to avoid forking this code
excessively.

Note that logging is enabled only in debug builds, and we force LOG() and
LOG_VERBOSE() to never be enabled. This will be changed in a future patch.

Bug: 5448972
Change-Id: I8b056743ae09adc518658a639185c3983dd56573