Initialize properties of JavaValue
When converting from NPVariant to JavaValue, there are cases where we don't do
an explicit conversion, but instead rely on the default value of the JavaValue
for the corresponding type. We need to make sure that the default values are
correctly set.

This bug was introduced in http://trac.webkit.org/changeset/82194.

Change-Id: I681a305e57a706d54e26c1e0d00ccc2d657bedd9