Fix Java Bridge when coercing JavaScript string to Java char in arrays
Currently, when converting coercing a JavaScript string to a method parameter
of type Java char, we always use a value of 0.

However, when coercing a JavaScript object to method parameter of type Java
char array, we behave differently. In this case, if the element is a JavaScript
string, we use the first character of the string for the char value.

This change modifies the behaviour in the array case to match that in the
non-array case.

Bug: 5619677
Change-Id: Ifeee8940af0bb01df58eeff6c9dd7e1b40688790