Fix the repaint inval mechanism for layers - cherry-pick from MR1
We only keep the scrollable layers invalidating their entire area
instead of incurring this penalty for *all* layers.

Also remove unnecessary repaint when scrolling (we have the entire
content on the UI side already).

While the entire scrollable area will be invalidated and marked
as dirty, the existing code in PaintedSurface already only look at
the visible tiles of scrollable layers to consider the layer's content
as being ready to display, so the real world penalty (while far from
optimal) is limited.

Implementing the correct approach (only invalidating what really
changed on scrollable layers) would sadly be a lot more complex,
as currently webkit will *not* send us the repaint invals if they
are on a currently clipped area, as webkit's default behaviour
to implement scrolling of such element is to repaint them anyway...

bug:5721618
Change-Id: I9af240d5f42f00f9ab07710e43ef8d72dfaf8155