Clean up all the old code paths before TransferQueue
TransferQueue is using one GL texture per tile, which is the best in terms of
memory usage.
The old code paths include EGLImage mode and SurfaceTexture per tile mode.
At the same time, all the locks related to EGLImage are deleted too.

There should be no funcationality change here.
By deleting the DoubleBufferTexture, the only thing need to be added into
BaseTileTexture is the TextureInfo, which can be optimized further in the
future.

Change-Id: If27fbca7cb55c73d94fbde3ba02d6208707ad8f0