Don't pass pictures with no content to UI thread
bug:5748182

Adjusted text detection logic to also simply detect content, and moved to a
separate class. No content means a picture is not attached to the layer, and a PaintedSurface isn't allocated to the layer on the UI side.

Most commonly, this avoids allocating textures for the clipping element for an
overscroll div, but only if it has nothing to paint.

Change-Id: I5e57d497780c0532e0f28ac94b7ced3de9718e92