Employ the transfer queue to the pure color tiles
This fixed the sync problem for the pure color tiles.
When Tex Gen thread detect that the tile is pure color, UI need to pick it up
in the right time. To avoid adding too many sync object, the simplest way is
following the same idea of transfer queue.
At the same time, using a separate queue and adding simple operation on top of
it can limit the impact to the original transfer queue which is highly coupled
with SurfTex implemenation.

bug:5808996

Change-Id: I518d358c5626815b27f285d4dcadcee9c3234835