Only draw cursor rings once.
 Bug: 5853891

Drawing GLExtras for each layer might have fixed MatchInfo positions, but has
had the side-effect of causing cursor rings to be drawn once per layer,
resulting in multiple cursor rings being drawn when links are clicked. Oops!
Making certain that cursor rings are only drawn when GLExtras are rendering
the correct layer addresses this issue.

This has a side-effect of rendering cursor rings in the correct position when
layers have been scrolled, which was not hapening before this change. Bonus!

Change-Id: Ie55e132dddd5b420b1738aa6049229996083008a