Do not share SkPicture* between WebViewCore and CachedRoot.
 Bug: 5879729

WebViewCore was in the rather naughty habit of forcefully deleting an
SkPicture pointer that was currently in use by CachedRoot on the UI
thread. It isn't even necessary for WebViewCore to hold this pointer
once it has been fed to the CachedRoot, so I have transferred ownership
to the CachedRoot itself, who instead of forcefully deleting the
SkPicture simply unref()s it when it is finished. This addresses the
nasty crashes we've been seeing when the UI thread attempts to do
something with the SkPicture that has been deleted behind its back by
the WebKit thread.

Change-Id: Ifb364f822fc03d2c3813b2edb2b038124c65fc1f