Remove option to build WebKit with JSC
V8 has been the default JS engine since Froyo and has shipped on all devices in
Gingerbread onwards. However, we continued to support JSC for a number of
reasons. See bug for details.

Now that we can use V8 in all cases, we should drop support for JSC to
eliminate the maintenance overhead.

Adds a clean step as this changes a Makefile.

Note that this was previously landed as
https://android-git.corp.google.com/g/#/c/162144/, but broke the full_x86-eng
build, so was reverted.

Bug: 5495373
Change-Id: I3d41362f93b9f3d034f17150aaa402379e75d367