Remove option to build WebKit with JSC
V8 has been the default JS engine since Froyo and has shipped on all devices in
Gingerbread onwards. However, we continued to support JSC because V8 requires
ARMv7 and at the time, the emulator did not support this.

Now that the emulator supports ARMv7, we should drop support for JSC to
eliminate the maintenance overhead.

Bug: 5495373
Change-Id: Ib31fa1b4927099cf16e9e465233c8667317e209c