Plumb WebView and WebViewCore thru to MediaLayer
Depends on https://android-git.corp.google.com/g/171224

Changes the plugin code to provide the WebView (not WebViewClassic) jobject
instance into the plugin, as this provides the full public API expected.
As a knock-on, I had to plumb WebViewCore jobject through to MediaTexture
so it can retain the sendPluginDrawMsg() call. This isn't pretty, but in
terms of layering it's no worse than before.

Change-Id: Iaaa6e0c6b65963fedaff078a20477d8595c82a5d