Prefer CpuUpload path on the emulator (DO NOT MERGE)
Lack of synchronization between filling a tile and consuming it cause
incorrect tile contents to be used. Theoretically this can happen on
any system, but has only been observed on the emulator (where it
happens regularly). So change the default tile upload path to
CpuUpload on the emulator only.

This is a low-risk workaround for ics-mr1. Proper synchronization will
be added in less risk-averse branches.

Change-Id: I9a5dc2fd50c66f1686a881245b14567a885dd6ec