Accept-Language header is not updated when phone language changes
The Java class WebSettings has accessor methods for User Agent
and Accept-Language strings. These methods ensure that the values
are updated if the Locale has changed. However, WebSettings.cpp
in WebKit reads the private fields in WebSettings.java directly,
and thereby fails to trigger the update.

The correction is to use the appropriate accessor methods from
WebSettings.cpp.

Change-Id: Id1aa1f4de38d2a73cad9660afc35944c663dbd4a