Cherry-pick WebKit change r84762 to fix URL canonicalization
This is to make sure URL parsing behaviour is up-to-date prior to fixing
window.location. See bug for details.

Note that the 'http:example.com/' test case in fast/url/standard-url.html fails
with DRT2 because of an assumption in the test that it is served as a file
scheme URL.

See http://trac.webkit.org/changeset/84762

Bug: 2159848
Change-Id: Ie680debcad3d285efd40134932b72d429c0d2222