Translucent background color support for webview
We need to respect the background color set for webview and combine that with
the CSS background.
Since we are tiled based and need to avoid double drawing, we need to determine
the areas not covered by available tiles and draw the background. This involves
a bit computation overhead. Therefore, we only compute those areas when alpha
is between 1 to 254.
When alpha is 255, we just need a glClear. When alpha is 0, nothing need to be
drawn.

Due to the fact that the tile can be partially drawn, texture generator need to
know the background color to clear the tiles for the base layer.

bug:5684832

Change-Id: I51f9d219f2722cc088eeea603623f9660f7a136a