Cherry-pick WebKit change r90068 to fix a rendering crash
Note that this change does not apply cleanly in RenderBlock::destroy() without
r85869.

r85869 was intended to be an optimization only, with no functional change.
However, it introduced two crashes, one of which is fixed in r90068 and the
other in r86060. To avoid the need to cherry-pick both r85869 and r86060 as
prerequisites, the patch for r90068 was applied manually to give the correct
end result in RenderBlock::destroy().

We now pass the LayoutTests added in r90068 and the LayoutTest added in r86060
does not crash.

See http://trac.webkit.org/changeset/90068

Bug: 5496903
Change-Id: Ica6dc2c70a4cd5c781ae77df95a8a4033c097708