Fix crash in RenderLayerCompositor::updateCompositingLayers
Don't enable composited layers for fixed elements with a width or height
of 1 or less. This duplicates the check used for deciding if layers
following fixed layers are composited or not.

This fixes the scenario where updateBacking() switches composited mode
on (by calling enableCompositingMode()) because a layer is fixed, while
computeCompositingRequirements() will disable it, and results in a
segfault in updateCompositingLayers()

Bug: 6324603
Change-Id: I034f2bd2f1a84fc24de6ab359c07f4895e2c7121