Workaround double blit failure
An issue occurs on Stingray where calling glTexSubImage2d in order to blit an
update to a texture fails silently if the texture has already been updated this
frame.

With this change, we now draw the tile off-screen to work around the issue just
before the second blit is required. This way, the second blit succeeds.

bug:6630040

Change-Id: I07213954669b0f3d3f5b87a748a591e532b097b4