Clean appcache properly on deleteAllData()
Bug: 6763306

The prior approach queries and deletes only the resources that are
given in the manifest file. There are two problems it seems:
1. It is expensive & complicated. The deletealldata is called by an API
method which clearly states this will clear all javascript storage APIs.
there is an webstorage API for this, why make it complicated?

2. It seems a bug somewhere actually prevents cleaning disk storage.
(see the bug above for experiments). This can be a bug of the old approach,
or this can be a webkit bug (less likely).

Change-Id: I7cf984efce5117191ca0e691007e2e0a11d75e93