approximate text bounds for recording context
Currently, SkPaint::measureText() is very, very slow. Fortunately, we
do not need precise text bounds and can live with approximated bounds
so long as the approximation is not smaller than the glyphs. We err
very much on the side of safety so as to play nicely with calligraphic
fonts and scripts that make heavy use of modifiers.

Change-Id: Id08f0856c0d902386089560fc3cfcdd853a06177