Use combined depth and stencil buffers
Use combined depth and stencil buffers when both are requested together.
When stencil and depth buffers are requested, create a packed buffer
instead of standalone buffers and attach it to both the DEPTH_ATTACHMENT
point as well as the STENCIL_ATTACHMENT point as per
OES_packed_depth_stencil extension to GLES2.

Change-Id: I88a8f6f2e7eaf7cc7c93d37c03a9f7777ccc2732
CRs-fixed: 45920